refactor(landing-screen): update import path for EmptyWarehouse component
