+++ frontend/components/warehouse/empty-warehouse.tsx
